All_in_one_Chatbot_Summarizer_and_chatbot/requirements.txt
streamlit
python-dotenv
langchain
PyPDF2
faiss-cpu
openai
tiktoken
tqdm
python-docx
langchain-community